Move AmbiguousDiffMinusCounter to hunk_header
